Add failing test for nested key_format! call
Since #486, key format is applied when a hash is merged into the
currently built object:

    json.key_format! :upcase
    json.merge!({some: "value"})

    # => {"SOME": "value"}

As a side effect of the change, key format is also applied to the
result of nested blocks [1]:

    json.key_format! camelize: :lower
    json.level_one do
      json.key_format! :upcase
      json.value 'two'
    end

    # => jbuilder 2.10.0: {"levelOne": {"VALUE": "two"}}
    # => jbuilder 2.11.0: {"levelOne": {"vALUE": "two"}}

The string "vALUE" results from calling
`"value".upcase.camelize(:lower)`.

This makes it impossible to override key format in child elements.

[1] https://github.com/blackjack26/jbuilder/blob/3c0c882f23e892dbe51be625bb71adc42febae09/lib/jbuilder.rb#L39